Added support for option injection on passing OUTPUT verification
At the moment, we are inserting only an IP option. TCP / UDP options to
follow later (giving user a choice between the three).

Note that the signature is not being calculated yet. Need to decide
fields that are hashed & how to manage the signing key. See [1] when
implementing this part.

[1] https://wiki.openssl.org/index.php/EVP_Signing_and_Verifying