Added suport for identifying mount namespaces
This is a first step toward adding support for filtering traffic to/from
docker containers.

This patch checks if the root filesystem of a process is using
overlayfs. If so, then it identifies the upper overlay from its
filesystem options and prepends the absolute path of the merged
directory to the path of any object identified in its /proc/<pid>/maps.

Note that this does not mean we have support for chroot-ed processes.
Adding support for this is not a priority right now.