Moved process pool / object hashes identificaiton in get_verdict()
In preparation for adding network namespaces as a mandatory match
criteria, we moved the ({src,dst}_ip, {src,dst}_port) -> pid pool
resolution in get_verdict() and are doing it for every rule evaluation.

The reason is that it's exceedingly hard to find a socket based on
the ip / port fields without knowning the namespace it belongs to.
This would entail yet another procfs walk, on a per-thread basis.

In stead, we decided to require the user to specify the namespace
to which each rule relates. The plan is to switch namespaces before each
rule evaluation and perform the netlink socket diagnostics query from
said namespace. If no result is found, we can't be held accountable for
the process running in yet another namespace.

TODO: Find a way to split traffic from the FORWARD chain into
      INPUT / OUTPUT (if concerning containers) or true FORWARD if our
      machine indeed acts as a middlebox (and only does signature
      verification). Maybe add 2 more iptables rules that take into
      account the source / destination (docker bridge) interfaces and
      redirect those packets to the DAF INPUT / OUTPUT chain, instead of
      the FORWARD one.

NOTE: The current implementation does absolutely nothing for
      memoization, so all hashes are racalculated for every rule
      evaluation. Something _could_ be done there, but at this point we
      have no pretense that this whole thing can run optimally in
      userspace.