Added net namespace cache & match criteria
On insert / append operations, the user must specify a netns magic file
path (e.g. "/proc/1/ns/net"). On the firewall side, before inserting the
rule on the required chain, the file is opened and a file descriptor is
also stored in the same rule struct.

Each magic file is opened exactly once, then cached. Note that while
there is even a single reference to this file (i.e. open file
descriptor) at the system level, the namespace will not be deleted.
However, chances are that it won't appear when doing a `lsns` after all
the processes that were part of it died.

On rule deletion, the refcounter is decremented and the cached file
descriptor is closed (if refcount reaches 0).