Now able to look up sockets in multiple network namespaces
For each check, we switch the network namespace to that which is
specified by the rule. The socket lookup slow path will thus yield the
desired socket. The socket chaching mechanism is now updated to take
into account the network namespace device & inode numbers, as well as
the bound port.

TODO: bypass setns() call if target ns is the same