Started work on multi-threaded implementation
This patch only separates the event handler dispatcher from the main()
function and adds the worker threads entry point. These are not yet
utilized.

Note that the plan at the moment is to create _one_ thread for each
epoll priority class, meaning one for system events and one for packet
processing. The support for extending this to multiple threads is there
(see wake conditions, for example) but chances are that doing so will
incur more work on synchronizing system state structure updates.