Added parallelized option (w/o data structure access protection)
This is not great at the moment. Using the default level-triggered
epollin behaviour floods the work queue with already pending events.
Switching to edge-triggered prevents some events from ever triggering
again. The solution here seems to be marking the leaf file descriptors
as oneshot, only to reactivate them in the worker thread once the
processing is done. This causes a significant increase in syscalls which
tanks the performance.

Personal note: Best to run iperf with `-w 131070`, otherwise the
               netfilter queue module will complain about insufficient
               buffer space. Take a look at the `-w` entry in the iperf
               man page. Note that altering the received packet buffer
               size in `nfq_set_mode` does nothing.