Added option for event priority uniformization
This places all events under a single epoll object, thus elliminating
the need for an epoll hierarchy. As a result, packet processing events
will have a higher chance of being selected for processing. The
improvement in throughput is marginal. Note that not stalling packet
processing this way creates the risk of evaluation with incomplete or
outdated information.