Added option for partial userpace packet copy from NFQ
The '-P' option will transfer only 80 bytes of data in userspace. This
should be enough to cover the IP header, any potential IP options and
the TCP header. Expect ~132 bytes of data to be effectively transferred.
This includes the netlink protocol header (16B) and an arbitrary amount
of padding.

Regardless, this should be a significant reduction from the roughly
MTU-sized packets that were being copied over. Hopefully, the slow
memory access speeds should no longer be an issue on the 10Gbps servers
with DDR3 RAM.

Note: consider expanding the option to allow 40B transfers, assuming
      that no IP options are ever used.