Added pwru output parsing script for packet procesising CFG analysis
This script takes pwru output (with relative timestamps) and generates a
set of Control Flow Graphs with average elapsed times for node
transitions. When the piped output terminates (via a Ctrl+C, for
example), the script will dump all unique paths to stdout.

Use this to find the origin of the packet processing delays, depending
on MTU (i.e.: spikes after 6k MTU).