Fixed issue with verdict batching timeout
Now the epoll operations in the main loop will timeout after the same
interval as the verdict forced transmission timeout. At the end of each
loop, the verdict transmission criteria are re-evaluated. Now, a
connection can no longer be stalled if no other traffic is being
generated by the system.