Replaced gettimeofday() with RDTSC
The former VDSO call incurred an overhead due to the timestamp counter
update lock. The coarse timer wouldn't have helped since it has 4ms
granularity and we need microsecond-level resolution.

Note that we require an invariable frequency timer (Fn8000_0007:EDX_8=1)
and the ability of using RDTSC from user space (CR4.TSD=0).

Make sure to set the BASE_FREQ variable. You can get the base frequency
from sysfs; just make sure that it's in Hz and not KHz:
    /sys/devices/system/cpu/cpu*/cpufreq/base_frequency

If this file is not exposed by the kernel, just run a simple program
measuring the elapsed cycles for a sleep(1).

NOTE: This implementation is _incorrect_ for older CPUs with variable
      frequency timers.

TODO: Maybe replace the gettimeofday() implementation in the elapsed
      time measurement macros from util.h.